Feat::UpdateCaregiverFacade

+++ src/main/java/pe/edu/upc/center/platform/caregiver/domain/model/commands/CreateCaregiverCommand.java
@@ -2,5 +2,5 @@ package pe.edu.upc.center.platform.caregiver.domain.model.commands;
 
 
 public record CreateCaregiverCommand(String completeName, Integer age, String address, Integer caregiverExperience, Integer completedServices,
-         String biography, String profileImage, Long farePerHour, String districtsScope) {
+                                     String biography, String profileImage, Double farePerHour, String districtsScope) {
 }
